docker-image-templates/image-resize-template/requirements.txt
requests
minio
enum
Pillow
boto3==1.10.9